refactor: move button and info content to components folder
